shortener-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
grpcio==1.64.0
grpcio-tools==1.64.0
pydantic==2.7.1
